Add and edit time trials, including adding multiple people
Add lots of people when creating a time trial
If you forgot someone, go back and add them in later

No way to remove people or time trials at the moment, except through the
console.

Using a mixture of turbo for adding extra people and stimulus for
initially adding in people. This is because you need a time trial id to
return the turbo form to add a new person, but with stimulus, it uses
the templated form object without having to do a server round trip